use SliderParameter in 2d view

+++ duality-client/src/duality/VolumeRenderer2D.h
@@ -3,21 +3,20 @@
 #include "IVDA/GLMatrix.h"
 #include "IVDA/Vectors.h"
 #include "duality/CoordinateSystem.h"
-#include "src/duality/VolumeDataset.h"
 #include "src/duality/TransferFunction.h"
+#include "src/duality/VolumeDataset.h"
 
-#include <memory>
 #include <array>
+#include <memory>
 
 class GLShader;
-
 class VolumeRenderer2D {
 public:
     VolumeRenderer2D();
     ~VolumeRenderer2D();
-    
-    void render(const VolumeDataset& dataset, const GLMatrix& mvp, const TransferFunction& tf, CoordinateAxis axis, float slice);
-    
+
+    void render(const VolumeDataset& dataset, const GLMatrix& mvp, const TransferFunction& tf, CoordinateAxis axis, int slice);
+
 private:
     std::unique_ptr<GLShader> m_shader;
     std::array<std::array<IVDA::Vec3f, 4>, 3> m_vertices;
